각 기능들 디테일 수정 및 test 완료.

+++ simpleproject/src/main/java/com/example/demo/user/UserRepository.java
@@ -8,6 +8,6 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
 	UserEntity findByUserId(String userId);
 
-	UserEntity findByUserNumAndUserId(int userNum, String userId);
+	UserEntity findByUserNumAndUserId(Long userNum, String userId);
 
 }
